refactor(dataverse): create card component
