+++ src/component/card/dataverseCard/dataverseCard.scss
@@ -1,0 +1,94 @@
+@import '@/style/fonts.scss';
+@import '@/style/mixins.scss';
+
+.okp4-dataverse-portal-dataverse-card-main {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  position: relative;
+
+  .okp4-dataverse-portal-tag-main {
+    position: absolute;
+    min-width: 137px;
+    max-height: 40px;
+    right: 0;
+    text-transform: uppercase;
+    text-align: center;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: pre-wrap;
+    /* stylelint-disable-next-line value-no-vendor-prefix */
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+  }
+
+  .okp4-dataverse-portal-dataverse-card-content {
+    width: 84%;
+    padding-top: 55px;
+    padding-bottom: 24px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 27px;
+
+    @include desktop-bp {
+      width: 88%;
+    }
+
+    .okp4-dataverse-portal-dataverse-description {
+      display: flex;
+      flex-direction: column;
+      width: 100%;
+      gap: 10px;
+      text-align: center;
+      font-size: 16px;
+
+      h3 {
+        @include noto-sans(700);
+        width: inherit;
+        line-height: 28px;
+        letter-spacing: 0.4px;
+        white-space: nowrap;
+        overflow: hidden;
+        text-overflow: ellipsis;
+      }
+
+      p {
+        @include noto-sans(300);
+        width: inherit;
+        overflow: hidden;
+        text-overflow: ellipsis;
+        white-space: pre-wrap;
+        height: 44px;
+        font-size: 16px;
+        line-height: 22px;
+        opacity: 0.8;
+        /* stylelint-disable-next-line value-no-vendor-prefix */
+        display: -webkit-box;
+        -webkit-line-clamp: 2;
+        -webkit-box-orient: vertical;
+
+        img {
+          height: inherit;
+          object-fit: cover;
+        }
+
+        a:visited {
+          color: themed(text);
+        }
+      }
+    }
+
+    .okp4-dataverse-portal-button-main {
+      display: flex;
+      justify-content: center;
+      width: 100%;
+
+      @include desktop-bp {
+        width: fit-content;
+      }
+    }
+  }
+}
