+++ src/page/dataverse/dataverse.scss
@@ -61,90 +61,6 @@
         grid-template-columns: repeat(auto-fit, minmax(330px, 1fr));
         gap: 29px;
       }
-
-      .okp4-dataverse-portal-dataverse-page-card {
-        width: 100%;
-        height: 100%;
-        display: flex;
-        justify-content: center;
-        position: relative;
-
-        .okp4-dataverse-portal-tag-main {
-          position: absolute;
-          width: 137px;
-          right: 0;
-          text-transform: uppercase;
-          text-align: center;
-        }
-
-        .okp4-dataverse-portal-dataverse-page-card-content {
-          width: 84%;
-          padding-top: 55px;
-          padding-bottom: 24px;
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-          gap: 27px;
-
-          @include desktop-bp {
-            width: 88%;
-          }
-
-          .okp4-dataverse-portal-dataverse-page-description {
-            display: flex;
-            flex-direction: column;
-            width: 100%;
-            gap: 10px;
-            text-align: center;
-            font-size: 16px;
-
-            h3 {
-              @include noto-sans(700);
-              width: inherit;
-              line-height: 28px;
-              letter-spacing: 0.4px;
-              white-space: nowrap;
-              overflow: hidden;
-              text-overflow: ellipsis;
-            }
-
-            p {
-              @include noto-sans(300);
-              width: inherit;
-              overflow: hidden;
-              text-overflow: ellipsis;
-              white-space: pre-wrap;
-              height: 44px;
-              font-size: 16px;
-              line-height: 22px;
-              opacity: 0.8;
-              /* stylelint-disable-next-line value-no-vendor-prefix */
-              display: -webkit-box;
-              -webkit-line-clamp: 2;
-              -webkit-box-orient: vertical;
-
-              img {
-                height: inherit;
-                object-fit: cover;
-              }
-
-              a:visited {
-                color: themed(text);
-              }
-            }
-          }
-
-          .okp4-dataverse-portal-button-main {
-            display: flex;
-            justify-content: center;
-            width: 100%;
-
-            @include desktop-bp {
-              width: fit-content;
-            }
-          }
-        }
-      }
     }
   }
 }
